requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.0
pydantic>=2.10.0
openai>=1.20.0
requests>=2.32.0
gymnasium>=1.0.0
stable-baselines3>=2.8.0
numpy>=2.1.0
pytest>=8.2.0
python-multipart>=0.0.10
starlette>=0.38.0
python-dotenv>=1.0.1
